requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
sqlalchemy>=1.4.23
alembic>=1.7.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
aiokafka>=0.7.2
asyncpg>=0.25.0
python-jose>=3.3.0
passlib>=1.7.4
python-multipart>=0.0.5
kafka-python>=2.0.2
email-validator>=2.0.0
bcrypt>=4.0.1
